# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API UHS Asbl",
        "description": "API REST complète avec gestion des activites de UHS Asbl",
        "contact": {
            "name": "Support API",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api-uhs.uhs-rdc.org",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/payements.getAllData": {
            "get": {
                "tags": [
                    "Payements"
                ],
                "summary": "Liste des Payements",
                "description": "Récupérer toutes les Payements avec leurs abonnés",
                "operationId": "e8b837bf5416f65773264dbea8699390",
                "responses": {
                    "200": {
                        "description": "Liste des Payements",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Facturation"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/payements.store": {
            "post": {
                "tags": [
                    "Payements"
                ],
                "summary": "Créer un paiement",
                "operationId": "ab0d5906d86756a3b8e1aed131bf97b3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "loan_amount",
                                    "paid_amount",
                                    "transaction_date",
                                    "account_id",
                                    "facture_id"
                                ],
                                "properties": {
                                    "loan_amount": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 1000
                                    },
                                    "paid_amount": {
                                        "type": "number",
                                        "format": "decimal",
                                        "example": 500
                                    },
                                    "transaction_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-09-27"
                                    },
                                    "account_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "facture_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Paiement créé avec succès"
                    },
                    "400": {
                        "description": "Erreur de validation"
                    },
                    "500": {
                        "description": "Erreur serveur lors de la création"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Facturation": {
                "title": "Facturation",
                "description": "Modèle de facturation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "abonne_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "addedBy": {
                        "type": "integer",
                        "example": 1
                    },
                    "mois": {
                        "type": "string",
                        "example": "09-2025"
                    },
                    "montant": {
                        "type": "number",
                        "format": "float",
                        "example": 150.75
                    },
                    "dette": {
                        "type": "number",
                        "format": "float",
                        "example": 50
                    },
                    "status": {
                        "type": "string",
                        "example": "Non payé"
                    },
                    "reference": {
                        "type": "string",
                        "example": "FAC-6637"
                    },
                    "date_emission": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-09-01"
                    },
                    "date_paiement": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-09-10",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-21T10:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-09-21T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}